ml-service/requirements.txt
flask==2.3.3
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2
python-dotenv==1.0.0
flask-cors==4.0.0
gunicorn==21.2.0